sankhya-agent/requirements.txt
mcp
requests
python-dotenv
streamlit
plotly
pandas
openai
pkgutil
importlib
inspect
re
logging
sqlite3
typing
